perf(showcase): keep the stream path off the default demo
The frozen `/` page no longer statically imports `@adapttable/core/stream`.
Realtime wiring lives behind the realtime page boundary, so a bundler
cannot pull the stream module into the live demo.
